Fixed Naming Conventions and Folder Structure
