+++ src/main/java/com/example/dashboardbackend/config/DataSeeder.java
@@ -5,14 +5,13 @@ import com.example.dashboardbackend.models.Family;
 import com.example.dashboardbackend.models.User;
 import com.example.dashboardbackend.models.enums.UserAvatarType;
 import com.example.dashboardbackend.models.enums.UserRole;
-import com.example.dashboardbackend.models.widgets.WidgetConfig;
-import com.example.dashboardbackend.models.widgets.WidgetItem;
-import com.example.dashboardbackend.models.widgets.WidgetPosition;
+import com.example.dashboardbackend.models.widgets.general.WidgetConfig;
+import com.example.dashboardbackend.models.widgets.general.WidgetItem;
+import com.example.dashboardbackend.models.widgets.general.WidgetPosition;
 import com.example.dashboardbackend.repositories.DashboardRepository;
 import com.example.dashboardbackend.repositories.FamilyRepository;
 import com.example.dashboardbackend.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
