+++ src/main/java/com/example/dashboardbackend/controllers/MediaController.java
@@ -1,9 +1,7 @@
 package com.example.dashboardbackend.controllers;
 
-import com.example.dashboardbackend.dtos.MediaUploadRequest;
 import com.example.dashboardbackend.services.MediaService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
