--- src/main/java/com/example/dashboardbackend/dtos/FamilyLoginResponse.java
@@ -1,10 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import java.util.List;
-
-public record FamilyLoginResponse(
-        String token,
-        List<UserProfile> profiles,
-        String type
-) {
-}
