--- src/main/java/com/example/dashboardbackend/dtos/UpdateWidgetConfigRequest.java
@@ -1,9 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import com.example.dashboardbackend.models.widgets.WidgetConfig;
-
-public record UpdateWidgetConfigRequest(
-        WidgetConfig widgetConfig
-) {
-
-}
