--- src/main/java/com/example/dashboardbackend/dtos/UpdateWidgetPositionRequest.java
@@ -1,8 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import com.example.dashboardbackend.models.widgets.WidgetPosition;
-
-public record UpdateWidgetPositionRequest(
-        WidgetPosition widgetPosition
-) {
-}
