--- src/main/java/com/example/dashboardbackend/dtos/UserSelectPageResponse.java
@@ -1,8 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import java.util.List;
-
-public record UserSelectPageResponse(
-        List<UserProfile> userProfiles
-) {
-}
