+++ src/main/java/com/example/dashboardbackend/dtos/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.example.dashboardbackend.dtos;
+package com.example.dashboardbackend.dtos.auth;
 
 public record LoginRequest(
         String name,
