+++ src/main/java/com/example/dashboardbackend/dtos/dashboard/DashboardResponse.java
@@ -6,6 +6,6 @@ public record DashboardResponse(
         Long id,
         Long familyId,
         List<WidgetResponse> widgets,
-        Permissions permissions
+        PermissionResponse permissionResponse
 ) {
 }
