+++ src/main/java/com/example/dashboardbackend/dtos/media/MediaUploadRequest.java
@@ -1,4 +1,4 @@
-package com.example.dashboardbackend.dtos;
+package com.example.dashboardbackend.dtos.media;
 
 import org.springframework.web.multipart.MultipartFile;
