+++ src/main/java/com/example/dashboardbackend/dtos/user/UserSelectPageResponse.java
@@ -1,0 +1,8 @@
+package com.example.dashboardbackend.dtos.user;
+
+import java.util.List;
+
+public record UserSelectPageResponse(
+        List<UserProfileResponse> userProfileResponses
+) {
+}
