+++ src/main/java/com/example/dashboardbackend/dtos/user/UserSelectRequest.java
@@ -1,4 +1,4 @@
-package com.example.dashboardbackend.dtos;
+package com.example.dashboardbackend.dtos.user;
 
 public record UserSelectRequest(
         Long userId,
