+++ src/main/java/com/example/dashboardbackend/dtos/weather/CurrentResponse.java
@@ -3,7 +3,7 @@ package com.example.dashboardbackend.dtos.weather;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record Current(
+public record CurrentResponse(
         @JsonProperty("temperature_2m") double temperature,
         @JsonProperty("wind_speed_10m") double windSpeed,
         @JsonProperty("weather_code") int weatherCode
