+++ src/main/java/com/example/dashboardbackend/dtos/weather/GeoLocationListResponse.java
@@ -1,0 +1,8 @@
+
+
+package com.example.dashboardbackend.dtos.weather;
+
+import java.util.List;
+
+public record GeoLocationListResponse(List<GeoLocationResponse> results) {
+}
