+++ src/main/java/com/example/dashboardbackend/dtos/widgets/UpdateWidgetConfigRequest.java
@@ -1,0 +1,9 @@
+package com.example.dashboardbackend.dtos.widgets;
+
+import com.example.dashboardbackend.models.widgets.general.WidgetConfig;
+
+public record UpdateWidgetConfigRequest(
+        WidgetConfig widgetConfig
+) {
+
+}
