+++ src/main/java/com/example/dashboardbackend/dtos/widgets/UpdateWidgetPositionRequest.java
@@ -1,0 +1,8 @@
+package com.example.dashboardbackend.dtos.widgets;
+
+import com.example.dashboardbackend.models.widgets.general.WidgetPosition;
+
+public record UpdateWidgetPositionRequest(
+        WidgetPosition widgetPosition
+) {
+}
