+++ src/main/java/com/example/dashboardbackend/models/widgets/weather/WeatherData.java
@@ -1,6 +1,6 @@
 package com.example.dashboardbackend.models.widgets.weather;
 
-import com.example.dashboardbackend.models.widgets.WidgetItem;
+import com.example.dashboardbackend.models.widgets.general.WidgetItem;
 import jakarta.persistence.Id;
 
 public class WeatherData {
