+++ src/main/java/com/example/dashboardbackend/repositories/WidgetItemRepository.java
@@ -1,10 +1,7 @@
 package com.example.dashboardbackend.repositories;
 
-import com.example.dashboardbackend.models.widgets.WidgetItem;
+import com.example.dashboardbackend.models.widgets.general.WidgetItem;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
 
 public interface WidgetItemRepository extends JpaRepository<WidgetItem, Long> {
 }
