+++ src/main/java/com/example/dashboardbackend/services/AuthenticationService.java
@@ -1,9 +1,10 @@
 package com.example.dashboardbackend.services;
 
-import com.example.dashboardbackend.dtos.*;
-import com.example.dashboardbackend.dtos.auth.AuthRequest;
-import com.example.dashboardbackend.dtos.auth.AuthResponse;
-import com.example.dashboardbackend.dtos.auth.RegisterRequest;
+import com.example.dashboardbackend.dtos.auth.*;
+import com.example.dashboardbackend.dtos.family.FamilyLoginResponse;
+import com.example.dashboardbackend.dtos.user.UserProfileResponse;
+import com.example.dashboardbackend.dtos.user.UserSelectRequest;
+import com.example.dashboardbackend.dtos.user.UserSelectResponse;
 import com.example.dashboardbackend.exceptions.FamilyNotFoundException;
 import com.example.dashboardbackend.models.Family;
 import com.example.dashboardbackend.models.enums.UserAvatarType;
@@ -92,7 +93,7 @@ public class AuthenticationService {
         else if (principal instanceof FamilyPrincipal familyPrincipal) {
             Family family = familyPrincipal.getFamily();
             String token = jwtUtils.generateFamilyToken(family);
-            List<UserProfile> profiles = getProfiles(family.getId());
+            List<UserProfileResponse> profiles = getProfiles(family.getId());
             return new FamilyLoginResponse(token, profiles, "FAMILY");
         }
 
@@ -123,9 +124,9 @@ public class AuthenticationService {
         return new UserSelectResponse(token, user.getId(), user.getUserRole());
     }
 
-    private List<UserProfile> getProfiles(Long familyId) {
+    private List<UserProfileResponse> getProfiles(Long familyId) {
         return userRepository.findByFamilyId(familyId).stream()
-                .map(user -> new UserProfile(
+                .map(user -> new UserProfileResponse(
                         user.getId(),
                         user.getName(),
                         user.getAvatar(),
